requirements.txt
# Requirements for Vercel deployment
# This file is used by Vercel to install dependencies

fastapi==0.104.1
uvicorn==0.24.0
openai>=1.12.0
pydantic==2.5.0
python-dotenv==1.0.0
mangum>=0.17.0
httpx>=0.28.1

